Improvement of code
- Small change in exampleAnalysis.m: ‘/‘ is missing in the folder path.

- Change in exampleImportData.m:  In case maxNumOfComp > size (X, 1),
we change the way we call function pcamat (original version).

- New function exampleImportDataFromFile.m: new function to load data
using a file to load path and IDs of the data, which can be located in
different folders.

- New function loadFromFile.m: Read the complete list of subjects paths
and IDs from a txt file and save them in the output array of cells.

- File path_subjects_ADHD_ExtDisk_subsetPekin.txt: example of txt file
with subject paths and IDs.

-  Small changes in examplePermutationTest.m: missing path of the
validation folder and definition of N, number of elements in Z.